GUI update + cohesion lvl increased by 100 lvls - divided studentclient.fxml and studentpanel.fxml and their respective controllers
